Fix image-copy-ecr example.
We share this example with customers a lot but it has grown a bit stale.

Mainly, the assumption that 'repository' will be equal to <org>/<repo>
is not true and is a source of confusion.

This change modifies the example to use the `repo_id` to look up the name of the
repo in the Chainguard API. The name of the organization is passed in by
the user when running the terraform.

Also:
- Remove demo instructions from the README. We don't provide write
  access to the registry so the instructions can't be followed.
- Bump the Go version and Go modules.

Signed-off-by: Rob Best <rob.best@chainguard.dev>