Update image-copy-gcp example
Modify the `image-copy-gcp` example to account for the fact that
registry push events do not contain the literal repository name any
more.

We use the `repo_id` in the event body to lookup the repository name in
the Chainguard API. The name of the organization is passed into the terraform
by the user.

I've also ported over some other improvements for the image-copy-ecr
example:
1. Copy the tag referred to in the event, rather than just copying
   latest every time.
2. Ignore failed pushes, events without tags and non-manifest pushes.
3. Store the image-copy container image in the repository that the
   terraform creates. This stops us from leaving the image hanging
   around in GAR when the terraform is destroyed.

Signed-off-by: Rob Best <rob.best@chainguard.dev>