image-copy-{ecr,gcp}: optionally ignore referrers (#238)
If customers aren't familiar with cosign then the sea of
`sha256-abcdef...` tags in the mirror repositories can be confusing and
make it unnecessarily difficult to understand the available tags,
especially if they have no intention of doing anything with the
signatures and attestations.

This change adds a variable (`ignore_referrers`) that disables the
mirroring of signatures and attestations.

I've updated the READMEs to document this variable.

I've also lengthened the timeout on the `image-copy-ecr` Lambda. The
default of 3 seconds is far too short and I was running into timeouts
while testing this change.