image-mapper: add matcher that identifies iamguarded images
Some iamguarded images have aliases for the upstream, non-bitnami
variants. But not all. To provide a more consistent mapping behaviour,
let's try and infer a match from upstream -> iamguarded.

Not everyone is going to be interested in using our Helm charts, so add
an option to ignore those matches.

As part of this I've made the ignore functionality more modular so it's
easier to add other ignore logic in future.