image-mapper: fix bad aliases in the data (#253)
There are images where the alias data is known to be incorrect and it is
non-trivial to fix on our side at the moment. For now, let's correct
those aliases in the mapper to improve the results.

Add an integration test that tests against the actual data returned by
GraphQL so we have a baseline for identifying regressions and codifying
known-good output.